# apps/api/openapi.json
{"openapi":"3.0.0","paths":{"/users":{"get":{"operationId":"UsersController_findAll","summary":"Get all users.","description":"Endpoint that returns a list of all users.","parameters":[],"responses":{"200":{"description":""}}},"post":{"operationId":"UsersController_create","summary":"Create a user.","description":"Endpoint that creates a new user.","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUsersDto"}}}},"responses":{"201":{"description":""}}}},"/users/{userId}":{"get":{"operationId":"UsersController_findOne","summary":"Get a user.","description":"Endpoint that returns a single user by their user id.","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FindOneUserDto"}}}}}},"delete":{"operationId":"UsersController_delete","summary":"Delete a user.","description":"Endpoint that deletes a user by their user id.","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}}}},"/users/{userId}/comms":{"get":{"operationId":"CommsController_findAll","summary":"Get all communications.","description":"Endpoint that returns a list of all communications or conversations for a user.","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/FindOneCommDto"}}}}}}},"post":{"operationId":"CommsController_create","summary":"Create a communication.","description":"Endpoint that creates a new communication or conversation for a user.","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCommsDto"}}}},"responses":{"201":{"description":""}}}},"/users/{userId}/comms/{commId}":{"get":{"operationId":"CommsController_findOne","summary":"Get a communication.","description":"Endpoint that returns a single communication or conversation for a user, by their user and communication id.","parameters":[{"name":"commId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FindOneCommDto"}}}}}},"delete":{"operationId":"CommsController_delete","summary":"Delete a communication.","description":"Endpoint that deletes a single communication or conversation for a user, by their user and communication id.","parameters":[{"name":"commId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}}}},"/users/{userId}/relationships":{"get":{"operationId":"RelationshipsController_","summary":"Get all relationships.","description":"Endpoint that returns a list of all relationships for a user.","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/FindOneRelationshipDto"}}}}}}},"post":{"operationId":"RelationshipsController_","summary":"Create a relationship.","description":"Endpoint that creates a new relationship for a user.","parameters":[],"responses":{"201":{"description":""}}}},"/users/{userId}/relationships/{relationshipId}":{"get":{"operationId":"RelationshipsController_","summary":"Get a relationship.","description":"Endpoint that returns a single relationship for a user, by their user and relationship id.","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FindOneRelationshipDto"}}}}}},"delete":{"operationId":"RelationshipsController_","summary":"Delete a relationship.","description":"Endpoint that deletes a single relationship for a user, by their user and relationship id.","parameters":[],"responses":{"200":{"description":""}}}},"/users/{userId}/prefs":{"get":{"operationId":"PrefsController_findAll","summary":"Get all preferences.","description":"Endpoint that returns a list of all preferences for a user.","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/FindOnePrefDto"}}}}}}},"post":{"operationId":"PrefsController_create","summary":"Create a preference.","description":"Endpoint that creates a new preference for a user.","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePrefsDto"}}}},"responses":{"201":{"description":""}}}},"/users/{userId}/prefs/{prefId}":{"get":{"operationId":"PrefsController_findOne","summary":"Get a preference.","description":"Endpoint that returns a single preference for a user, by their user and preference id.","parameters":[{"name":"prefId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FindOnePrefDto"}}}}}},"delete":{"operationId":"PrefsController_delete","summary":"Delete a preference.","description":"Endpoint that deletes a single preference for a user, by their user and preference id.","parameters":[{"name":"prefId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}}}},"/embeddings/o":{"get":{"operationId":"EmbeddingsController_query","parameters":[{"name":"query","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}}}},"/embeddings":{"post":{"operationId":"EmbeddingsController_generateEmbeddings","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenerateEmbeddingsDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenerateEmbeddingsResponseDto"}}}}}}}},"info":{"title":"Oktavolt API Specification","description":"","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"FindOneUserDto":{"type":"object","properties":{"userId":{"type":"number"},"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"phone":{"type":"string"},"dateOfBirth":{"format":"date-time","type":"string"},"createdAt":{"format":"date-time","type":"string"}},"required":["userId","firstName","lastName","email","phone","dateOfBirth","createdAt"]},"CreateUserDto":{"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"phone":{"type":"string"},"dateOfBirth":{"format":"date-time","type":"string"},"context":{"type":"string"}},"required":["firstName","lastName","email","phone","dateOfBirth"]},"CreateUsersDto":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/CreateUserDto"}}},"required":["data"]},"FindOneCommDto":{"type":"object","properties":{"commId":{"type":"number"},"type":{"type":"object"},"content":{"type":"string"},"timestamp":{"format":"date-time","type":"string","nullable":true},"sender":{"type":"string","nullable":true},"receiver":{"type":"string"},"provider":{"type":"object"}},"required":["commId","type","content","timestamp","sender","receiver","provider"]},"CreateCommDto":{"type":"object","properties":{"type":{"type":"object"},"content":{"type":"string"},"timestamp":{"format":"date-time","type":"string"},"sender":{"type":"string"},"receiver":{"type":"string"},"provider":{"type":"object"}},"required":["type","content","receiver","provider"]},"CreateCommsDto":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/CreateCommDto"}}},"required":["data"]},"FindOneRelationshipDto":{"type":"object","properties":{"relationshipId":{"type":"number"},"name":{"type":"string"},"relationType":{"type":"object"},"email":{"type":"string","nullable":true},"phone":{"type":"string","nullable":true},"address":{"type":"string","nullable":true},"notes":{"type":"string","nullable":true}},"required":["relationshipId","name","relationType","email","phone","address","notes"]},"FindOnePrefDto":{"type":"object","properties":{"prefId":{"type":"number"},"userId":{"type":"number"},"preferenceType":{"type":"string"},"value":{"type":"string","nullable":true}},"required":["prefId","userId","preferenceType","value"]},"CreatePrefDto":{"type":"object","properties":{"preferenceType":{"type":"string"},"value":{"type":"string"},"context":{"type":"string"}},"required":["preferenceType"]},"CreatePrefsDto":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/CreatePrefDto"}}},"required":["data"]},"GenerateEmbeddingsDto":{"type":"object","properties":{"content":{"type":"string"}},"required":["content"]},"GenerateEmbeddingsResponseDto":{"type":"object","properties":{"embeddings":{"type":"array","items":{"required":true,"type":"array","items":{"type":"number"}}}},"required":["embeddings"]}}}}